rag-engine/requirements.txt
# Core dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Document processing
PyPDF2>=3.0.1
python-docx>=1.0.0
unstructured>=0.10.30
email-validator>=2.0.0

# Vector database
pinecone-client>=2.2.4

# LLM and embeddings
google-generativeai>=0.3.1
openai>=1.3.0
tiktoken>=0.5.1

# Utilities
python-multipart>=0.0.6
logging>=0.4.9.6
asyncio>=3.4.3